first test iteration

+++ src/main/kotlin/com/onemorepet/models/PetOffer.kt
@@ -1,3 +1,3 @@
 package com.onemorepet.models
 
-data class PetOffer(val kind: String, val species: List<Pair<String, Float>>, val age: UInt, val price: UInt)
+data class PetOffer(val kind: String, val species: List<Pair<String, Double>>, val age: Int, val price: Int)
